+++ src/main/kotlin/io/github/erayerdin/kappdirs/appdirs/AppDirs.kt
@@ -15,7 +15,7 @@ interface AppDirs {
      * For more information, see [here](https://kappdirs.readthedocs.io/en/latest/windows-system/#what-is-roaming).
      * @return A directory containing data related to a specific user.
      */
-    fun getUserDataDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean = false): Path
+    fun getUserDataDir(appName: String, appVersion: String, appAuthor: String? = null, roaming: Boolean = false): Path
 
     /**
      * @param appName The name of the application.
@@ -26,7 +26,7 @@ interface AppDirs {
      * For more information, see [here](https://kappdirs.readthedocs.io/en/latest/windows-system/#what-is-roaming).
      * @return A directory containing configuration data of the application related to a specific user.
      */
-    fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean = false): Path
+    fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String? = null, roaming: Boolean = false): Path
 
     /**
      * @param appName The name of the application.
@@ -35,7 +35,7 @@ interface AppDirs {
      * It does not have any effect on Mac OS X, any Unix-based systems or any Linux distributions.
      * @return A directory containing cache data of the application. It is a user-based directory.
      */
-    fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String): Path
+    fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String? = null): Path
 
     /**
      * @param appName The name of the application.
@@ -44,7 +44,7 @@ interface AppDirs {
      * It does not have any effect on Mac OS X, any Unix-based systems or any Linux distributions.
      * @return A directory containing log files of the application. It is a user-based directory.
      */
-    fun getUserLogDir(appName: String, appVersion: String, appAuthor: String): Path
+    fun getUserLogDir(appName: String, appVersion: String, appAuthor: String? = null): Path
 
     /**
      * @param appName The name of the application.
@@ -55,7 +55,7 @@ interface AppDirs {
      * Linux distributions.
      * @return A directory containing system-wide data of the application.
      */
-    fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String, local: Boolean = false): Path
+    fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String? = null, local: Boolean = false): Path
 
     /**
      * @param appName The name of the application.
@@ -64,5 +64,5 @@ interface AppDirs {
      * It does not have any effect on Mac OS X, any Unix-based systems or any Linux distributions.
      * @return A directory containing system-wide configuration files of the application.
      */
-    fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String): Path
+    fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String? = null): Path
 }