+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+ - Moved `AppDirs` interface from package root to `appdirs` package
+
+### Removed
+ - `local` argument on `AppDirs::getSiteConfigDir`
+
 ## [0.1.4-prealpha] - 2019-03-16
 ### Added
  - Docstrings